requirements.txt
win10toast==0.8
python-Levenshtein==0.12.0